feat: 재고 수정 화면 연동
